tests:Refactor|| Refactored Session Varaibles to main Scope and Clean Package.
