+++ tests/crud/conftest.py
@@ -1,33 +1,13 @@
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import Session
-from sqlalchemy.pool import StaticPool
-
 from semesterstat.common.reports import ScoreReport, StudentReport, SubjectReport
-from semesterstat.database import BatchSchemeInfo, Department, Score, Student, Subject
-from semesterstat.database.models import Base
+from semesterstat.database import Score, Student, Subject
+from sqlalchemy.orm import Session
 
 
-def input_data(_db: Session):
-    depts = [
-        {"Code": code, "Name": name}
-        for (code, name) in [
-            ("CS", "Computer Science"),
-            ("IS", "Information Science"),
-            ("TE", "Telecommunication"),
-            ("ME", "Mechanical Engineering"),
-            ("AE", "Aeronautical Engineering"),
-        ]
-    ]
+@pytest.fixture(scope="package", autouse=True)
+def input_data(engine):
 
-    _db.bulk_insert_mappings(Department, depts)
-
-    batch_scheme = [
-        {"Batch": batch, "Scheme": scheme}
-        for (batch, scheme) in [(2015, 2015), (2016, 2015), (2017, 2017)]
-    ]
-
-    _db.bulk_insert_mappings(BatchSchemeInfo, batch_scheme)
+    db = Session(bind=engine)
 
     stu_data = [
         StudentReport(Usn=usn, Name=name)
@@ -39,7 +19,7 @@ def input_data(_db: Session):
         ]
     ]
     student = [x.dict() for x in stu_data]
-    _db.bulk_insert_mappings(Student, student)
+    db.bulk_insert_mappings(Student, student)
 
     subect_data = [
         SubjectReport(
@@ -55,7 +35,7 @@ def input_data(_db: Session):
         ]
     ]
     subjects = [x.dict() for x in subect_data]
-    _db.bulk_insert_mappings(Subject, subjects)
+    db.bulk_insert_mappings(Subject, subjects)
 
     score_data = [
         ScoreReport(
@@ -71,46 +51,7 @@ def input_data(_db: Session):
         ]
     ]
     scores = [x.dict() for x in score_data]
-    _db.bulk_insert_mappings(Score, scores)
+    db.bulk_insert_mappings(Score, scores)
 
-    _db.commit()
-
-
-@pytest.fixture(scope="session")
-def engine():
-    engine = create_engine(
-        "sqlite://", connect_args={"check_same_thread": False}, poolclass=StaticPool
-    )
-
-    Base.metadata.create_all(bind=engine)
-
-    input_data(Session(bind=engine))
-
-    # Return the Session Data.
-    yield engine
-
-    engine.dispose()
-
-
-@pytest.fixture(scope="function")
-def db(engine):
-    _engine = engine
-
-    # Make a new Connection.
-    conn = _engine.connect()
-
-    # OverArching Transaction, doesn't matter what happened before.
-    trans = conn.begin()
-
-    session = Session(bind=conn)
-
-    yield session
-
-    # Dispose the objects in memory, we don't need them.
-    session.close()
-
-    # Rollback all Commits.
-    trans.rollback()
-
-    # Close the connection
-    conn.close()
+    db.commit()
+    db.close()
